# WebAPI_Kurs_KW48_ppedv/WebAPICourseService/bin/Debug/net5.0/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebAPICourseService",
    "version": "v1"
  },
  "paths": {
...
    "/api/Movie": {
      "get": {
        "tags": [
          "Movie"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              },
              "text/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              },
              "text/csv": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Movie"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          }
        }
      }
    },
    "/api/Movie/{id}": {
      "get": {
        "tags": [
          "Movie"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Movie"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Movie"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "GenreTyp": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "format": "int32"
      },
      "Movie": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "genre": {
            "$ref": "#/components/schemas/GenreTyp"
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}